+++ ErsatzTV.Core/Domain/MediaItem/ImageFolderDuration.cs
@@ -5,5 +5,5 @@ public class ImageFolderDuration
     public int Id { get; set; }
     public int LibraryFolderId { get; set; }
     public LibraryFolder LibraryFolder { get; set; }
-    public int DurationSeconds { get; set; }
+    public double DurationSeconds { get; set; }
 }
